+++ next-app/src/components/LayoutSwitcher/LayoutSwicher.view.tsx
