+++ next-app/src/components/ReportEditor/ReportEditor.view.tsx
