+++ next-app/src/components/ReportPreview/ReportPreview.view.tsx
