+++ next-app/src/domain/model/report.ts
@@ -1,4 +1,4 @@
-import { User } from '@/app/_domain/model/user'
+import { User } from '@/domain/model/user'
 
 interface IReport {
   id: number
