+++ next-app/src/domain/model/user.ts
